@@ -1,29 +1,29 @@
 {
   "name": "wfs-spy-react",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.15",
-    "@types/react": "^17.0.39",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
     "@types/react-bootstrap-typeahead": "^5.1.8",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.3.3",
     "@types/reactstrap": "^8.7.1",
-    "axios": "^0.25.0",
-    "bootstrap": "^4.6.0",
+    "axios": "^1.2.2",
+    "bootstrap": "^5.2.3",
     "font-awesome": "^4.7.0",
-    "highcharts": "^9.3.3",
+    "highcharts": "^10.3.2",
     "highcharts-react-official": "^3.1.0",
-    "node-sass": "^7.0.1",
-    "react": "^17.0.2",
-    "react-bootstrap-typeahead": "^5.2.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "react-toastify": "^8.1.1",
-    "reactstrap": "^8.9.0",
-    "typescript": "^4.5.5"
+    "node-sass": "^8.0.0",
+    "react": "^18.2.0",
+    "react-bootstrap-typeahead": "^6.0.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.1",
+    "react-scripts": "^5.0.1",
+    "react-toastify": "^9.1.1",
+    "reactstrap": "^9.1.5",
+    "typescript": "^4.9.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -39,7 +39,10 @@
     "collectCoverageFrom": [
       "src/**/*.{ts,tsx}",
       "!src/index.tsx"
-    ]
+    ],
+    "moduleNameMapper": {
+      "axios": "axios/dist/node/axios.cjs"
+    }
   },
   "browserslist": {
     "production": [
@@ -54,27 +57,26 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
-    "eslint": "^8.8.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
+    "eslint": "^8.30.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest-dom": "^4.0.1",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "eslint-plugin-testing-library": "^5.0.5",
-    "husky": "^7.0.4",
-    "jest-watch-typeahead": "^0.6.5",
-    "lint-staged": "^12.3.3",
-    "msw": "^0.36.8",
-    "prettier": "^2.5.1"
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest-dom": "^4.0.3",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^5.9.1",
+    "husky": "^8.0.2",
+    "jest-watch-typeahead": "^2.2.1",
+    "lint-staged": "^13.1.0",
+    "msw": "^0.49.2",
+    "prettier": "^2.8.1"
   },
   "husky": {
     "hooks": {
