@@ -6,7 +6,7 @@ const totalItems: string = "Total Items:";
 
 export default function TotalItems({ numberOfItems }: TotalItemsProps) {
   return (
-    <div className="float-right mb-2">
+    <div className="float-end mb-2">
       {totalItems} <b>{numberOfItems}</b>
     </div>
   );
