@@ -1,25 +1,1 @@
-import { Col } from "reactstrap";
-
-import { Panel } from "../../../components";
-import FeatureDetails from "./FeatureDetails";
-import { useWindowWidth } from "../../../hooks";
-import { getSizeAndOffset } from "../../../utils";
-import consts from "./constants";
-
-export default function FeatureTypeList() {
-  const windowWidth = useWindowWidth();
-  const { size, offset } = getSizeAndOffset(windowWidth);
-
-  return (
-    <Col md={{ size, offset }} className="mt-4">
-      <h3>{consts.header}</h3>
-      <p>{consts.descr}</p>
-
-      <Panel
-        content={<FeatureDetails />}
-        header={consts.cardHeader}
-        title={consts.cardTitle}
-      />
-    </Col>
-  );
-}
+export { default } from "./FeatureTypeList";
