@@ -1,5 +1,4 @@
-import { fireEvent } from "@testing-library/react";
-import { act, renderHook } from "@testing-library/react-hooks";
+import { fireEvent, act, renderHook } from "@testing-library/react";
 
 import useWindowWidth from "./useWindowWidth";
 
