@@ -1,10 +1,8 @@
 import getTimeInMs from "./getTimeInMs";
 
 test("retuns input in miliseconds", () => {
-  const mockDate = getTimeInMs();
-  const spy = jest
-    .spyOn(global, "Date")
-    .mockImplementation(() => mockDate as unknown as string);
+  const timeInMs = getTimeInMs();
+  const dateNowSpy = jest.spyOn(Date, "now").mockImplementation(() => timeInMs);
 
-  spy.mockRestore();
+  dateNowSpy.mockRestore();
 });
